Security fix for work command

+++ src/commands/shop/roles/modules/cancel.ts
@@ -3,30 +3,30 @@ import {
   CommandInteraction,
   ColorResolvable,
   GuildMemberRoleManager,
-} from "discord.js";
+} from 'discord.js';
 
 // Configurations
-import config from "../../../../../config.json";
+import config from '../../../../../config.json';
 
 // Models
-import userSchema from "../../../../helpers/database/models/userSchema";
-import shopRolesSchema from "../../../../helpers/database/models/shopRolesSchema";
+import shopRolesSchema from '../../../../helpers/database/models/shopRolesSchema';
 
 // Helpers
-import creditNoun from "../../../../helpers/creditNoun";
+import creditNoun from '../../../../helpers/creditNoun';
+import fetchUser from '../../../../helpers/fetchUser';
 
 // Function
 export default async (interaction: CommandInteraction) => {
   const { options, guild, user, member } = interaction;
 
-  const optionRole = options.getRole("role");
+  const optionRole = options.getRole('role');
 
   // If amount is null
   if (optionRole === null) {
     // Embed object
     const embed = {
-      title: ":dollar: Shop - Roles [Cancel]" as string,
-      description: "We could not read your requested role." as string,
+      title: ':dollar: Shop - Roles [Cancel]' as string,
+      description: 'We could not read your requested role.' as string,
       color: config?.colors?.error as ColorResolvable,
       timestamp: new Date() as Date,
       footer: {
@@ -51,10 +51,9 @@ export default async (interaction: CommandInteraction) => {
     await guild?.roles
       .delete(optionRole?.id, `${user?.id} canceled from shop`)
       .then(async () => {
-        const userDB = await userSchema?.findOne({
-          userId: user?.id,
-          guildId: guild?.id,
-        });
+        const userDB = await fetchUser(user, guild);
+
+        if (userDB === null) return;
 
         await shopRolesSchema?.deleteOne({
           roleId: optionRole?.id,
@@ -63,12 +62,12 @@ export default async (interaction: CommandInteraction) => {
         });
 
         const embed = {
-          title: ":shopping_cart: Shop - Roles [Cancel]" as string,
+          title: ':shopping_cart: Shop - Roles [Cancel]' as string,
           description: `You have canceled ${optionRole.name}.` as string,
           color: config?.colors?.success as ColorResolvable,
           fields: [
             {
-              name: "Your balance" as string,
+              name: 'Your balance' as string,
               value: `${creditNoun(userDB?.credits)}` as string,
             },
           ],
